Updated Dockerfile to prepare it for app deployment
First, the Command for the container execution has been add. It simply runs our
run.sh file, where the actual code to run our application server is.
Second, the new /scripts (./scripts locally) directory has been add, for
helper scripts that will help with the deployment.

Note: A suggested dependency was not added, because I am not sure is
even necessary for the current docker image.